+++ Source/PlayerClient/Public/PlayerClientGetOwnerAssetsBodyInput.h
@@ -49,7 +49,7 @@ public:
 	TOptional<SortByEnum> SortBy;
 	TOptional<bool> IncludeAttributes;
 	TOptional<FString> Continuation;
-	TOptional<double> ChainId;
+	TOptional<int64> ChainId;
 	TOptional<double> Limit;
 };
